requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Document parsing
pypdf==3.17.4
python-docx==1.1.0
Pillow==10.1.0

# OCR (optional, can be disabled)
easyocr==1.7.0
pytesseract==0.3.10

# Embeddings and vector store
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3

# Database (fallback for vector store)
aiosqlite==0.19.0

# Observability
prometheus-client==0.19.0
structlog==23.2.0

# LLM (optional, with dev stub mode)
openai==1.3.7
anthropic==0.7.7

# Utilities
python-dotenv==1.0.0
hashlib-compat==1.0.1
aiofiles==23.2.1
diff-match-patch==20230430

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
mypy==1.7.1

